requirements.txt
Django==4.2.15
djangorestframework==3.15.2
drf_yasg==1.21.7
geopy==2.4.1
Requests==2.32.3
